ci: ready for test
