+++ scripts/update_docusarus_config.js
@@ -62,36 +62,6 @@ const findApiDocsPluginVersionsObject = () => {
   return apiVersionsObject;
 };
 
-// Function to add noIndex: true after trailingSlash
-const addNoIndexProperty = () => {
-  // Find the main "config" declaration
-  const configDeclaration = ast.program.body.find(
-    (node) => node.type === "VariableDeclaration" && node.declarations[0].id.name === "config"
-  );
-
-  // If the "config" declaration is not found, log an error and return
-  if (!configDeclaration) {
-    console.error('Could not locate the main "config" declaration.');
-    return;
-  }
-
-  // Find the "trailingSlash" property in the "config" declaration
-  const trailingSlashProperty = configDeclaration.declarations[0].init.properties.find(
-    (prop) => prop.key.name === "trailingSlash"
-  );
-
-  // If "trailingSlash" is found, insert "noIndex: true" after it
-  if (trailingSlashProperty) {
-    const noIndexProperty = t.objectProperty(t.identifier("noIndex"), t.booleanLiteral(true));
-
-    // Find the index of the trailingSlash property and insert noIndex after it
-    const index = configDeclaration.declarations[0].init.properties.indexOf(trailingSlashProperty);
-    configDeclaration.declarations[0].init.properties.splice(index + 1, 0, noIndexProperty);
-  } else {
-    console.error('Could not locate the "trailingSlash" property.');
-  }
-};
-
 // This function takes in a "versionsObject" parameter and updates it with new properties based on the "versionsArray" and "versionsOverride" arrays
 const updateVersionsObject = (versionsObject) => {
   // Loop through each version in the "versionsArray"
@@ -165,8 +135,6 @@ const apiDocsVersionsObject = findApiDocsPluginVersionsObject();
 if (apiDocsVersionsObject) {
   updateVersionsObject(apiDocsVersionsObject);
 }
-
-addNoIndexProperty();
 
 // This is where the new config object is converted back into code.
 const updatedCode = generate(ast).code;
